pyproject.toml
[project]
name = "ai-codebench"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Aegeaner", email = "xihuke@gmail.com"},
]
dependencies = [
    "openai>=1.51.0",
    "anthropic>=0.34.0",
    "google-generativeai>=0.8.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "aiofiles>=24.1.0",
    "pyyaml>=6.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
